Updated Data Store ERD
Made changes to the Calendar table. Instead of a single owner_id to
identify the owner of the calendar, I split it into user_id and group_id
because a singluar owner_id wasn't going to work as a forgien key. Also
got rid of the user_or_group flag.